Reference

List of view types

With that as an introduction, let's list all the view types supported by this package. channelview and colorview are opposite transformations, as are rawview and ufixedview. channelview and colorview typically create objects of type ChannelView and ColorView, respectively, unless they are "undoing" a previous view of the opposite type.

channelview
ChannelView
colorview
ColorView
rawview
ufixedview
permuteddimsview

List of traits

ImageCore.pixelspacing — Function.

pixelspacing(img) -> (sx, sy, ...)

Return a tuple representing the separation between adjacent pixels along each axis of the image. Defaults to (1,1,...). Use ImagesAxes for images with anisotropic spacing or to encode the spacing using physical units.

source

ImageCore.spacedirections — Function.

spacedirections(img) -> (axis1, axis2, ...)

Return a tuple-of-tuples, each axis[i] representing the displacement vector between adjacent pixels along spatial axis i of the image array, relative to some external coordinate system ("physical coordinates").

By default this is computed from pixelspacing, but you can set this manually using ImagesMeta.

source

ImageCore.sdims — Function.

sdims(img)

Return the number of spatial dimensions in the image. Defaults to the same as ndims, but with ImagesAxes you can specify that some axes correspond to other quantities (e.g., time) and thus not included by sdims.

source

ImageCore.coords_spatial — Function.

coords_spatial(img)

Return a tuple listing the spatial dimensions of img.

Note that a better strategy may be to use ImagesAxes and take slices along the time axis.

source

ImageCore.size_spatial — Function.

size_spatial(img)

Return a tuple listing the sizes of the spatial dimensions of the image. Defaults to the same as size, but using ImagesAxes you can mark some axes as being non-spatial.

source

ImageCore.indices_spatial — Function.

indices_spatial(img)

Return a tuple with the indices of the spatial dimensions of the image. Defaults to the same as indices, but using ImagesAxes you can mark some axes as being non-spatial.

source

ImageCore.nimages — Function.

nimages(img)

Return the number of time-points in the image array. Defaults to 1. Use ImagesAxes if you want to use an explicit time dimension.

source

ImageCore.assert_timedim_last — Function.

assert_timedim_last(img)

Throw an error if the image has a time dimension that is not the last dimension.

source